+++ packages/web/app/features/Invoice/EditInvoiceModal/index.jsx
@@ -1,7 +1,6 @@
 import React, { useState } from 'react';
 import styled, { css } from 'styled-components';
 import { v4 as uuidv4 } from 'uuid';
-import * as yup from 'yup';
 import { Box, CircularProgress, Divider } from '@material-ui/core';
 import PrintIcon from '@material-ui/icons/Print';
 import { Form, Button, FormSubmitCancelRow, Modal, TranslatedText } from '@tamanu/ui-components';
@@ -20,6 +19,7 @@ import { InvoiceRecordModal } from '../../../components/PatientPrinting/modals/I
 import { PaymentTablesGroup } from './PaymentTablesGroup';
 import { useAuth } from '../../../contexts/Auth';
 import { NoteModalActionBlocker } from '../../../components/NoteModalActionBlocker';
+import { invoiceFormSchema } from '../InvoiceForm/invoiceFormSchema';
 
 const LinkText = styled.div`
   font-weight: 500;
@@ -105,107 +105,6 @@ export const EditInvoiceModal = ({
   const handleShowErrorDialog = errors => {
     return Object.keys(errors).length === 1 && errors['totalInsurerPercentage'];
   };
-
-  const schema = yup.object({
-    invoiceItems: yup.array(
-      yup.object().shape(
-        {
-          orderDate: yup.string().when(['productId', 'orderedByUserId'], {
-            is: (productId, orderedByUserId) => productId || orderedByUserId,
-            then: yup
-              .string()
-              .required(
-                <TranslatedText
-                  stringId="validation.required.inline"
-                  fallback="*Required"
-                  data-testid="translatedtext-8g9w"
-                />,
-              ),
-            otherwise: yup.string(),
-          }),
-          productId: yup.string().when(['orderDate', 'orderedByUserId'], {
-            is: (orderDate, orderedByUserId) => orderDate || orderedByUserId,
-            then: yup
-              .string()
-              .required(
-                <TranslatedText
-                  stringId="validation.required.inline"
-                  fallback="*Required"
-                  data-testid="translatedtext-wff4"
-                />,
-              ),
-            otherwise: yup.string(),
-          }),
-          orderedByUserId: yup.string().when(['orderDate', 'productId'], {
-            is: (orderDate, productId) => orderDate || productId,
-            then: yup
-              .string()
-              .required(
-                <TranslatedText
-                  stringId="validation.required.inline"
-                  fallback="*Required"
-                  data-testid="translatedtext-dz1y"
-                />,
-              ),
-            otherwise: yup.string(),
-          }),
-          quantity: yup
-            .number()
-            .required(
-              <TranslatedText
-                stringId="general.required"
-                fallback="Required"
-                data-testid="translatedtext-029d"
-              />,
-            ),
-          productPrice: yup.number(),
-        },
-        [
-          ['orderDate', 'productId'],
-          ['productId', 'orderedByUserId'],
-          ['orderDate', 'orderedByUserId'],
-        ],
-      ),
-    ),
-    insurers: yup.array(
-      yup.object({
-        insurerId: yup
-          .string()
-          .required()
-          .translatedLabel(
-            <TranslatedText
-              stringId="invoice.modal.editInvoice.insurer.label"
-              fallback="Insurer"
-              data-testid="translatedtext-ufad"
-            />,
-          ),
-        percentage: yup
-          .number()
-          .required(
-            <TranslatedText
-              stringId="general.required"
-              fallback="Required"
-              data-testid="translatedtext-vh20"
-            />,
-          ),
-      }),
-    ),
-    totalInsurerPercentage: yup
-      .mixed()
-      .test(
-        'totalInsurerPercentage',
-        <TranslatedText
-          stringId="invoice.modal.editInvoice.insurer.totalPercentageError"
-          fallback="Total insurer percentage must be less than or equal to 100%"
-          data-testid="translatedtext-ddnm"
-        />,
-        function(_, context) {
-          return (
-            context.parent.insurers.reduce((acc, curr) => acc + curr.percentage || 0, 0) <= 100
-          );
-        },
-      ),
-  });
 
   const renderDataTables = (values, formArrayMethods) => {
     if (editable) {
@@ -350,7 +249,7 @@ export const EditInvoiceModal = ({
                 }))
               : [],
           }}
-          validationSchema={schema}
+          validationSchema={invoiceFormSchema}
           render={({ submitForm, values }) => (
             <FieldArray name="invoiceItems" data-testid="fieldarray-3xyn">
               {formArrayMethods => {
